pyproject.toml
[project]
name = "mcp-server-odoo"
version = "0.1.0"
description = "MCP server for Odoo ERP integration - enables AI assistants to interact with Odoo"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Viktor Zeman", email = "vzeman@gmail.com"}]
license = {text = "MIT"}
keywords = ["mcp", "odoo", "erp", "ai", "assistant", "integration"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Financial :: Accounting",
]

dependencies = [
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/mcp-server-odoo"
"Bug Tracker" = "https://github.com/yourusername/mcp-server-odoo/issues"
"Documentation" = "https://github.com/yourusername/mcp-server-odoo#readme"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.hatch.build.targets.wheel]
packages = ["mcp_server_odoo"]